requirements.txt
numpy>=2.1
pandas>=2.2
scikit-learn>=1.5
matplotlib>=3.9
seaborn>=0.13
yfinance>=0.2.38
joblib>=1.3
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
pytz>=2024.1
streamlit>=1.38
streamlit-autorefresh>=0.1.3
